requirements.txt
Django==5.2.5
pillow==11.3.0
face-recognition==1.3.0
numpy==1.24.3
dlib==19.24.2
cmake==3.27.2
opencv-python==4.8.0.74